server/requirements.txt
fastapi[standard]
requests

sqlmodel
psycopg2-binary

bcrypt
pyjwt
cryptography

fastapi-cache2[redis]

pytest
httpx